Рефакторинг

+++ src/main/java/com/example/mywebquizengine/Demo.java
@@ -1,19 +1,24 @@
 package com.example.mywebquizengine;
 
+import com.example.mywebquizengine.user.model.domain.User;
 import com.example.mywebquizengine.user.repository.UserRepository;
+import org.springframework.amqp.core.Binding;
+import org.springframework.amqp.core.FanoutExchange;
+import org.springframework.amqp.core.Queue;
 import org.springframework.amqp.rabbit.core.RabbitAdmin;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
 
 import javax.transaction.Transactional;
+import java.util.List;
 
 
 @Component
 public class Demo implements CommandLineRunner {
 
     @Autowired
-    private UserRepository userRepository ;
+    private UserRepository userRepository;
 
     @Autowired
     private RabbitAdmin rabbitAdmin;
@@ -22,35 +27,34 @@ public class Demo implements CommandLineRunner {
     @Override
     public void run(String... args) {
 
-        /*List<User> users2 = userRepository.findAll();
+        boolean enabled = false;
 
-        for (User user : users2) {
+        if (!enabled) {
+            return;
+        }
+
+        List<User> users = userRepository.findAll();
+        for (User user : users) {
             rabbitAdmin.deleteQueue(user.getUsername());
+        }
 
-        }*/
+        if (rabbitAdmin.getQueueProperties("application") == null) {
 
-        //if (rabbitAdmin.getQueueProperties("application") == null) {
-
-          /*  List<User> users = userRepository.findAll();
-
-            for (User user : users) {
-
+            List<User> userList = userRepository.findAll();
+            for (User user : userList) {
                 for (int i = 0; i < user.getPhotos().size(); i++) {
                     user.getPhotos().get(i).setPosition(i);
-                }*/
-                //rabbitAdmin.declareExchange(new FanoutExchange(user.getUsername(), true, false));
-                /*Queue queue = new Queue(user.getUsername(), true, false, false);
+                }
+                rabbitAdmin.declareExchange(new FanoutExchange(user.getUsername(), true, false));
+                Queue queue = new Queue(user.getUsername(), true, false, false);
 
                 Binding binding = new Binding(user.getUsername(), Binding.DestinationType.QUEUE,
                         "message-exchange", user.getUsername(), null);
 
                 rabbitAdmin.declareQueue(queue);
-                rabbitAdmin.declareBinding(binding);*/
-
-
-           // }
-        //}
-//
-
+                rabbitAdmin.declareBinding(binding);
+            }
+        }
     }
+
 }
